building swift replacements for gmu objective c

+++ f4s-workexperience/KSPoint.swift
@@ -1,9 +1,9 @@
 
 public struct KSPoint: XYLocatable, Hashable, Equatable {
     public static let zero = KSPoint(x: 0, y: 0)
-    public var x: Float
-    public var y: Float
-    public init(x: Float, y: Float) {
+    public var x: Double
+    public var y: Double
+    public init(x: Double, y: Double) {
         self.x = x
         self.y = y
     }
