+++ f4s-workexperience/KSSize.swift
@@ -1,14 +1,14 @@
 
 public struct KSSize: Hashable, Equatable {
     public static let zero = KSSize(width: 0, height: 0)
-    public var width: Float
-    public var height: Float
-    public init(width: Float, height: Float) {
+    public var width: Double
+    public var height: Double
+    public init(width: Double, height: Double) {
         self.width = width
         self.height = height
     }
     
-    public func scaled(by f: Float) -> KSSize {
+    public func scaled(by f: Double) -> KSSize {
         return KSSize(width: width * f, height: height * f)
     }
 }
